refactor: types and dir str
