+++ packages/ecc-client-lit-ga4gh-tes/package.json
@@ -9,10 +9,7 @@
   "types": "./dist/index.d.ts",
   "componentsPrefix": "ecc-client-lit-ga4gh-tes-",
   "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    },
+    ".": { "types": "./dist/index.d.ts", "import": "./dist/index.js" },
     "./dist/custom-elements.json": "./dist/custom-elements.json",
     "./dist/index.js": "./dist/index.js",
     "./dist/components/*": "./dist/components/*",
@@ -33,10 +30,7 @@
     "lint:fix": "npm run lint -- --fix",
     "prepublish": "npm run build"
   },
-  "dependencies": {
-    "@elixir-cloud/design": "*",
-    "lit": "^2.8.0"
-  },
+  "dependencies": { "@elixir-cloud/design": "*", "lit": "^2.8.0" },
   "devDependencies": {
     "@custom-elements-manifest/analyzer": "^0.4.17",
     "@open-wc/eslint-config": "^9.2.1",
