+++ packages/ecc-client-lit-ga4gh-tes/src/components/create-run/create-runs.styles.ts
@@ -1,0 +1,5 @@
+import { css } from "lit";
+
+const styles = css``;
+
+export default styles;
