+++ packages/ecc-client-lit-ga4gh-tes/src/components/index.ts
@@ -1,2 +1,2 @@
-import "./tes-create-run/create-run.js";
+import "./create-run/create-run.js";
 import "./runs/runs.js";
