+++ packages/ecc-client-lit-ga4gh-tes/src/components/runs/runs.ts
@@ -1,78 +1,15 @@
 /* eslint-disable lit/no-classfield-shadowing */
 import { html, css, LitElement, render } from "lit";
 import { property, state } from "lit/decorators.js";
-import "@elixir-cloud/design";
+import {
+  FilterProp,
+  ItemProp,
+} from "@elixir-cloud/design/src/components/collection/index.js";
+import {
+  Field,
+  Action,
+} from "@elixir-cloud/design/src/components/details/index.js";
 import { deleteTask, fetchTask, fetchTasks } from "../../API/Task/tesGet.js";
-
-// TODO: Remove the interfaces once design package
-//       can export them, as they are copied from
-//       from there.
-
-export interface Children {
-  label?: string;
-  path: string;
-  copy?: boolean;
-  defaultValue?: any;
-}
-
-export interface Field {
-  key: string;
-  path: string;
-  tab?: string;
-  label?: string;
-  arrayOptions?: {
-    labelOptions?: {
-      path?: string;
-      prefix?: string;
-      suffix?: string;
-    };
-    type?: "detail" | "tag";
-  };
-  tooltip?: string;
-  copy?: boolean;
-  parentKey?: string;
-}
-export interface ItemProp {
-  index: number;
-  name: string;
-  key: string;
-  lazy?: boolean;
-  tag?: {
-    name: string;
-    type?: "primary" | "success" | "neutral" | "warning" | "danger";
-  };
-}
-
-export interface FilterProp {
-  key: string;
-  type: "search" | "select";
-  options?: string[];
-  selectConfig?: {
-    multiple?: boolean;
-  };
-  placeholder?: string;
-}
-
-export interface Action {
-  key: string;
-  label: string;
-  type: "button" | "link";
-  buttonOptions?: {
-    variant?: "primary" | "success" | "neutral" | "warning" | "danger" | "text";
-    loading?: boolean;
-    disabled?: boolean;
-    size?: "small" | "medium" | "large";
-    icon?: {
-      url: string;
-      position?: "prefix" | "suffix";
-    };
-  };
-  linkOptions?: {
-    url: string;
-    size?: "small" | "medium" | "large";
-  };
-  position?: "left" | "right";
-}
 
 export default class ECCClientGa4ghTesRuns extends LitElement {
   static styles = css``;
