+++ packages/ecc-client-lit-ga4gh-wes/src/components/create-run/create-run.ts
@@ -1,46 +1,7 @@
 import { html, LitElement } from "lit";
 import { property, state } from "lit/decorators.js";
+import { Field } from "@elixir-cloud/design/src/components/form/index.js";
 import { postWorkflow } from "../../API/Workflow/wesGet.js";
-import "@elixir-cloud/design";
-
-// TODO: import the interface from the design package
-export interface Field {
-  key: string;
-  label: string;
-  type?:
-    | "text"
-    | "date"
-    | "number"
-    | "email"
-    | "password"
-    | "tel"
-    | "url"
-    | "search"
-    | "datetime-local"
-    | "time"
-    | "array"
-    | "switch"
-    | "file"
-    | "group";
-  fieldOptions?: {
-    required?: boolean;
-    default?: string | boolean;
-    multiple?: boolean;
-    accept?: string;
-    returnIfEmpty?: string;
-    tooltip?: string;
-  };
-  arrayOptions?: {
-    defaultInstances?: number;
-    max?: number;
-    min?: number;
-  };
-  groupOptions?: {
-    collapsible: boolean;
-  };
-  error?: string;
-  children?: Array<Field>;
-}
 
 /**
  * @summary This component is used to create task runs using WES API.
@@ -55,7 +16,7 @@ export default class ECCClientGa4ghWesCreateRuns extends LitElement {
   @property({ type: String }) private baseURL =
     "https://prowes.rahtiapp.fi/ga4gh/wes/v1";
 
-  @state() fields: Array<Field> = [
+  @state() fields: Field[] = [
     {
       key: "workflow_url",
       label: "URL",
