+++ packages/ecc-client-lit-ga4gh-wes/src/components/index.ts
@@ -1,2 +1,2 @@
-import "./wes-create-run/wes-create-run.js";
+import "./create-run/create-run.js";
 import "./runs/runs.js";
